# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Buc", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/audit-assignments/": {
      "get": {
        "tags": ["audit-assignments"],
        "summary": "Read All Audit Assignments",
        "description": "Retrieve all audit assignments (Superuser/Admin only).",
        "operationId": "audit-assignments-read_all_audit_assignments",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuditAssignmentsPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["audit-assignments"],
        "summary": "Create Audit Assignment",
        "description": "Create new audit assignment from a template for a company or area.",
        "operationId": "audit-assignments-create_audit_assignment",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuditAssignmentCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuditAssignmentPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit-assignments/my-assignments": {
      "get": {
        "tags": ["audit-assignments"],
        "summary": "Read My Audit Assignments",
        "description": "Retrieve audit assignments that the current auditor can respond to.",
        "operationId": "audit-assignments-read_my_audit_assignments",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuditAssignmentsPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit-assignments/company/{company_id}": {
      "get": {
        "tags": ["audit-assignments"],
        "summary": "Read Audit Assignments For Company",
        "description": "Retrieve audit assignments for a specific company.",
        "operationId": "audit-assignments-read_audit_assignments_for_company",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Company Id" }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuditAssignmentsPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit-assignments/{assignment_id}": {
      "get": {
        "tags": ["audit-assignments"],
        "summary": "Read Audit Assignment By Id",
        "description": "Get a specific audit assignment by id.",
        "operationId": "audit-assignments-read_audit_assignment_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Assignment Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuditAssignmentPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["audit-assignments"],
        "summary": "Update Audit Assignment",
        "description": "Update an audit assignment.",
        "operationId": "audit-assignments-update_audit_assignment",
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Assignment Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuditAssignmentUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuditAssignmentPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["audit-assignments"],
        "summary": "Delete Audit Assignment",
        "description": "Delete an audit assignment.",
        "operationId": "audit-assignments-delete_audit_assignment",
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Assignment Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignedQuestionPublic": {
        "properties": {
          "text": { "type": "string", "title": "Text" },
          "question_type": { "$ref": "#/components/schemas/QuestionType" },
          "options": { "anyOf": [{}, { "type": "null" }], "title": "Options" },
          "order": { "type": "integer", "title": "Order" },
          "is_mandatory": { "type": "boolean", "title": "Is Mandatory" },
          "section_id": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Section Id"
          },
          "scoring_weight": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Scoring Weight"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "audit_assignment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Audit Assignment Id"
          },
          "original_question_template_id": {
            "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }],
            "title": "Original Question Template Id"
          }
        },
        "type": "object",
        "required": ["text", "question_type", "order", "is_mandatory", "id", "audit_assignment_id"],
        "title": "AssignedQuestionPublic"
      },
...
      "AuditAssignmentCreate": {
        "properties": {
          "title": { "type": "string", "maxLength": 255, "title": "Title" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "due_date": {
            "anyOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }],
            "title": "Due Date"
          },
          "status": { "$ref": "#/components/schemas/AuditAssignmentStatus", "default": "PENDING" },
          "periodicity": {
            "anyOf": [{ "$ref": "#/components/schemas/AuditPeriodicity" }, { "type": "null" }]
          },
          "next_due_date": {
            "anyOf": [{ "type": "string", "format": "date" }, { "type": "null" }],
            "title": "Next Due Date"
          },
          "is_public": { "type": "boolean", "title": "Is Public", "default": false },
          "audit_template_id": { "type": "string", "format": "uuid", "title": "Audit Template Id" },
          "company_id": { "type": "string", "format": "uuid", "title": "Company Id" },
          "area_id": {
            "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }],
            "title": "Area Id"
          }
        },
        "type": "object",
        "required": ["title", "audit_template_id", "company_id"],
        "title": "AuditAssignmentCreate"
      },
      "AuditAssignmentPublic": {
        "properties": {
          "title": { "type": "string", "maxLength": 255, "title": "Title" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "due_date": {
            "anyOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }],
            "title": "Due Date"
          },
          "status": { "$ref": "#/components/schemas/AuditAssignmentStatus", "default": "PENDING" },
          "periodicity": {
            "anyOf": [{ "$ref": "#/components/schemas/AuditPeriodicity" }, { "type": "null" }]
          },
          "next_due_date": {
            "anyOf": [{ "type": "string", "format": "date" }, { "type": "null" }],
            "title": "Next Due Date"
          },
          "is_public": { "type": "boolean", "title": "Is Public", "default": false },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "audit_template_id": { "type": "string", "format": "uuid", "title": "Audit Template Id" },
          "company_id": { "type": "string", "format": "uuid", "title": "Company Id" },
          "area_id": {
            "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }],
            "title": "Area Id"
          },
          "created_by_id": { "type": "string", "format": "uuid", "title": "Created By Id" },
          "created_at": { "type": "string", "format": "date-time", "title": "Created At" },
          "updated_at": { "type": "string", "format": "date-time", "title": "Updated At" },
          "audit_template": {
            "anyOf": [{ "$ref": "#/components/schemas/AuditTemplatePublic" }, { "type": "null" }]
          },
          "assigned_questions": {
            "items": { "$ref": "#/components/schemas/AssignedQuestionPublic" },
            "type": "array",
            "title": "Assigned Questions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "id",
          "audit_template_id",
          "company_id",
          "created_by_id",
          "created_at",
          "updated_at"
        ],
        "title": "AuditAssignmentPublic"
      },
      "AuditAssignmentStatus": {
        "type": "string",
        "enum": ["PENDING", "IN_PROGRESS", "COMPLETED", "OVERDUE"],
        "title": "AuditAssignmentStatus"
      },
      "AuditAssignmentUpdate": {
        "properties": {
          "title": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Title" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "due_date": {
            "anyOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }],
            "title": "Due Date"
          },
          "status": {
            "anyOf": [{ "$ref": "#/components/schemas/AuditAssignmentStatus" }, { "type": "null" }]
          },
          "periodicity": {
            "anyOf": [{ "$ref": "#/components/schemas/AuditPeriodicity" }, { "type": "null" }]
          },
          "next_due_date": {
            "anyOf": [{ "type": "string", "format": "date" }, { "type": "null" }],
            "title": "Next Due Date"
          },
          "area_id": {
            "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }],
            "title": "Area Id"
          },
          "is_public": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Public"
          }
        },
        "type": "object",
        "title": "AuditAssignmentUpdate"
      },
      "AuditAssignmentsPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/AuditAssignmentPublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "AuditAssignmentsPublic"
      },
      "AuditPeriodicity": {
        "type": "string",
        "enum": ["ONE_TIME", "DAILY", "WEEKLY", "MONTHLY", "QUARTERLY", "ANNUALLY"],
        "title": "AuditPeriodicity"
      },
...
      "AuditTemplatePublic": {
        "properties": {
          "name": { "type": "string", "maxLength": 255, "title": "Name" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "audit_type_definition_key": {
            "type": "string",
            "maxLength": 255,
            "title": "Audit Type Definition Key"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "created_by_id": { "type": "string", "format": "uuid", "title": "Created By Id" },
          "created_at": { "type": "string", "format": "date-time", "title": "Created At" },
          "updated_at": { "type": "string", "format": "date-time", "title": "Updated At" },
          "question_templates": {
            "items": { "$ref": "#/components/schemas/QuestionTemplatePublic" },
            "type": "array",
            "title": "Question Templates",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "audit_type_definition_key",
          "id",
          "created_by_id",
          "created_at",
          "updated_at"
        ],
        "title": "AuditTemplatePublic"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": { "message": { "type": "string", "title": "Message" } },
        "type": "object",
        "required": ["message"],
        "title": "Message"
      },
...
      "QuestionTemplatePublic": {
        "properties": {
          "text": { "type": "string", "title": "Text" },
          "question_type": { "$ref": "#/components/schemas/QuestionType" },
          "options": { "anyOf": [{}, { "type": "null" }], "title": "Options" },
          "order": { "type": "integer", "title": "Order" },
          "is_mandatory": { "type": "boolean", "title": "Is Mandatory", "default": true },
          "section_id": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Section Id"
          },
          "scoring_weight": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Scoring Weight"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "audit_template_id": { "type": "string", "format": "uuid", "title": "Audit Template Id" }
        },
        "type": "object",
        "required": ["text", "question_type", "order", "id", "audit_template_id"],
        "title": "QuestionTemplatePublic"
      },
...
      "QuestionType": {
        "type": "string",
        "enum": [
          "TEXT",
          "MULTIPLE_CHOICE_SINGLE",
          "MULTIPLE_CHOICE_MULTIPLE",
          "YES_NO",
          "RATING_SCALE",
          "SECTION_HEADER"
        ],
        "title": "QuestionType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" } }
      }
    }
  }
}